Cleanup test_runs
